+++ language-bindings/python/src/wamr/wamrapi/wamr.py
@@ -6,6 +6,7 @@ from ctypes import addressof
 from ctypes import c_char
 from ctypes import c_uint
 from ctypes import c_uint8
+from ctypes import c_uint64
 from ctypes import c_void_p
 from ctypes import cast
 from ctypes import create_string_buffer
@@ -167,7 +168,7 @@ class Instance:
             raise Exception("Error while creating module instance")
         return module_inst
 
-    def malloc(self, nbytes: int, native_handler) -> c_uint:
+    def malloc(self, nbytes: int, native_handler) -> c_uint64:
         return wasm_runtime_module_malloc(self.module_inst, nbytes, native_handler)
 
     def free(self, wasm_handler) -> None:
